Refactor: Consolidate driver connection and dispatch registries
