+++ sockets/dispatchRegistry.js
@@ -5,6 +5,40 @@ const dispatchedBookingToDriver = new Map();
 const DEFAULT_TTL_MS = 24 * 60 * 60 * 1000;
 const DISPATCH_TTL_MS = Number.parseInt(process.env.DISPATCH_TTL_MS || `${DEFAULT_TTL_MS}`, 10);
 
+// Runtime registry for connected drivers and their socket-level availability
+// Structure: driverId -> { socketIds: Set<string>, availableSockets: Set<string> }
+const driverConnectionRegistry = new Map();
+function ensureDriverEntry(driverId) {
+  const id = String(driverId);
+  if (!driverConnectionRegistry.has(id)) {
+    driverConnectionRegistry.set(id, { socketIds: new Set(), availableSockets: new Set() });
+  }
+  return driverConnectionRegistry.get(id);
+}
+function registerSocket(driverId, socketId) {
+  const entry = ensureDriverEntry(driverId);
+  entry.socketIds.add(String(socketId));
+}
+function unregisterSocket(driverId, socketId) {
+  const id = String(driverId);
+  const entry = driverConnectionRegistry.get(id);
+  if (!entry) return;
+  entry.socketIds.delete(String(socketId));
+  entry.availableSockets.delete(String(socketId));
+  if (entry.socketIds.size === 0 && entry.availableSockets.size === 0) {
+    driverConnectionRegistry.delete(id);
+  }
+}
+function setSocketAvailability(driverId, socketId, available) {
+  const entry = ensureDriverEntry(driverId);
+  const sid = String(socketId);
+  if (available) entry.availableSockets.add(sid); else entry.availableSockets.delete(sid);
+}
+function isDriverAvailableBySocket(driverId) {
+  const entry = driverConnectionRegistry.get(String(driverId));
+  return !!(entry && entry.availableSockets && entry.availableSockets.size > 0);
+}
+
 function makeKey(bookingId, driverId) {
   return `${String(bookingId)}:${String(driverId)}`;
 }
@@ -36,5 +70,12 @@ function cleanupDispatches() {
 
 setInterval(cleanupDispatches, DISPATCH_TTL_MS).unref();
 
-module.exports = { markDispatched, wasDispatched };
+module.exports = {
+  markDispatched,
+  wasDispatched,
+  registerSocket,
+  unregisterSocket,
+  setSocketAvailability,
+  isDriverAvailableBySocket
+};
